+++ nekoyume/game.py
@@ -6,7 +6,9 @@ from flask import (Blueprint, Response, g, redirect, render_template, request,
 from flask_babel import Babel
 from sqlalchemy import func
 
-from .models import LevelUp, Move, Node, User, cache, db
+from .models import LevelUp, Move, User, cache
+from .node import Node
+from .orm import db
 
 
 game = Blueprint('game', __name__, template_folder='templates')
